+++ ImagEditor2/src/effects/OutlineEffect.java
@@ -122,7 +122,7 @@ public class OutlineEffect extends Effect{
 			
 			@Override
 			public void actionPerformed(ActionEvent e) {
-				OutlineEffect.this.accuracy = accuracySlider.getValue();
+				OutlineEffect.this.accuracy = (int) accuracySlider.getValue();
 				OutlineEffect.this.outlineRGB = outlineColorLabel.getBackground().getRGB();
 				OutlineEffect.this.objectRGB = objectColorLabel.getBackground().getRGB();
 				parent.invalidate();
